# server/swagger/swagger.yml
basePath: /api/v2
consumes:
    - application/json
definitions:
...
    Board:
        description: Board groups a set of blocks and its layout
        x-go-package: github.com/antimatterchat/antimatter-plugin-boards/server/model
...
    BoardMember:
        description: BoardMember stores the information of the membership of a user on a board
        x-go-package: github.com/antimatterchat/antimatter-plugin-boards/server/model
...
    BoardPatch:
        description: BoardPatch is a patch for modify boards
        x-go-package: github.com/antimatterchat/antimatter-plugin-boards/server/model
...
    ErrorResponse:
        description: ErrorResponse is an error response
        x-go-package: github.com/antimatterchat/antimatter-plugin-boards/server/model
...
host: localhost
info:
    contact:
        email: api@focalboard.com
        name: Focalboard
        url: https://www.focalboard.com
    description: Focalboard Server
    license:
        name: Custom
        url: https://github.com/mattermost/focalboard/blob/main/LICENSE.txt
    title: Focalboard Server
    version: 2.0.0
paths:
...
    /boards/{boardID}:
        delete:
            description: Removes a board
            operationId: deleteBoard
            parameters:
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: success
                "404":
                    description: board not found
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
        get:
            description: Returns a board
            operationId: getBoard
            parameters:
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: success
                    schema:
                        $ref: '#/definitions/Board'
                "404":
                    description: board not found
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
        patch:
            description: Partially updates a board
            operationId: patchBoard
            parameters:
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
                - description: board patch to apply
                  in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/BoardPatch'
            produces:
                - application/json
            responses:
                "200":
                    description: success
                    schema:
                        $ref: '#/definitions/Board'
                "404":
                    description: board not found
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
...
    /boards/{boardID}/join:
        post:
            description: Become a member of a board
            operationId: joinBoard
            parameters:
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: success
                    schema:
                        $ref: '#/definitions/BoardMember'
                "403":
                    description: access denied
                "404":
                    description: board not found
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
    /boards/{boardID}/leave:
        post:
            description: Remove your own membership from a board
            operationId: leaveBoard
            parameters:
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: success
                "403":
                    description: access denied
                "404":
                    description: board not found
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
    /boards/{boardID}/members:
        get:
            description: Returns the members of the board
            operationId: getMembersForBoard
            parameters:
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: success
                    schema:
                        items:
                            $ref: '#/definitions/BoardMember'
                        type: array
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
        post:
            description: Adds a new member to a board
            operationId: addMember
            parameters:
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
                - description: membership to replace the current one with
                  in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/BoardMember'
            produces:
                - application/json
            responses:
                "200":
                    description: success
                    schema:
                        $ref: '#/definitions/BoardMember'
                "404":
                    description: board not found
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
    /boards/{boardID}/members/{userID}:
        delete:
            description: Deletes a member from a board
            operationId: deleteMember
            parameters:
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
                - description: User ID
                  in: path
                  name: userID
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: success
                "404":
                    description: board not found
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
        put:
            description: Updates a board member
            operationId: updateMember
            parameters:
                - description: Board ID
                  in: path
                  name: boardID
                  required: true
                  type: string
                - description: User ID
                  in: path
                  name: userID
                  required: true
                  type: string
                - description: membership to replace the current one with
                  in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/BoardMember'
            produces:
                - application/json
            responses:
                "200":
                    description: success
                    schema:
                        $ref: '#/definitions/BoardMember'
                default:
                    description: internal error
                    schema:
                        $ref: '#/definitions/ErrorResponse'
            security:
                - BearerAuth: []
...
produces:
    - application/json
schemes:
    - http
    - https
securityDefinitions:
    BearerAuth:
        description: 'Pass session token using Bearer authentication, e.g. set header "Authorization: Bearer <session token>"'
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"